test(react/checkbox): refactor checkbox tests to fit react v19
#231

re #231
