+++ packages/react/__test-utils__/render.tsx
@@ -1,0 +1,6 @@
+/** 請參考 https://testing-library.com/docs/react-testing-library/api#renderhook-options */
+export const createWrapper = (Element: any, props: any) => {
+  return function CreatedWrapper({ children }: { children: any }) {
+    return <Element {...props}>{children}</Element>;
+  };
+};
