+++ packages/react/src/Checkbox/CheckboxGroup.spec.tsx
@@ -1,20 +1,28 @@
-import { useContext, useState } from 'react';
-import {
-  cleanup,
-  fireEvent,
-  render,
-  renderHook,
-  TestRenderer,
-} from '../../__test-utils__';
+import { ReactNode, useContext, useState } from 'react';
+import { cleanup, fireEvent, render, renderHook } from '../../__test-utils__';
 import {
   describeForwardRefToHTMLElement,
   describeHostElementClassNameAppendable,
 } from '../../__test-utils__/common';
+import { createWrapper } from '../../__test-utils__/render';
 import {
   CheckboxGroupContext,
   CheckboxGroupContextValue,
 } from './CheckboxGroupContext';
 import Checkbox, { CheckboxGroup, CheckboxGroupOption } from '.';
+
+const renderMockCheckbox = jest.fn();
+
+jest.mock('./Checkbox', () => {
+  return function MockCheckbox(props: any) {
+    renderMockCheckbox(props);
+    return (
+      <div {...props}>
+        <input type="checkbox" checked={props.checked} />
+      </div>
+    );
+  };
+});
 
 describe('<CheckboxGroup />', () => {
   afterEach(cleanup);
@@ -46,13 +54,9 @@ describe('<CheckboxGroup />', () => {
       size: 'small',
       value: ['bar'],
     };
-    const { result, rerender } = renderHook(
-      () => useContext(CheckboxGroupContext),
-      {
-        wrapper: CheckboxGroup as any,
-        initialProps: expectProps,
-      },
-    );
+    const { result } = renderHook(() => useContext(CheckboxGroupContext), {
+      wrapper: createWrapper(CheckboxGroup, expectProps),
+    });
 
     /**
      * Ignore onChange since it will be transformed by CheckboxGroup
@@ -65,20 +69,12 @@ describe('<CheckboxGroup />', () => {
     }
 
     testCheckboxGroupContextValue();
-
-    expectProps = {
-      disabled: false,
-      name: 'bar',
-      size: 'large',
-      value: ['zoo'],
-    };
-    rerender(expectProps);
-    testCheckboxGroupContextValue();
   });
 
   describe('prop: options', () => {
     const options: CheckboxGroupOption[] = [
       {
+        disabled: false,
         label: 'foo',
         value: 'foo',
       },
@@ -89,18 +85,23 @@ describe('<CheckboxGroup />', () => {
       },
     ];
 
+    beforeEach(() => {
+      renderMockCheckbox.mockClear();
+    });
+
     it('should render checkboxes via options', () => {
-      const testInstance = TestRenderer.create(
-        <CheckboxGroup options={options} />,
-      );
-      const checkboxes = testInstance.root.findAllByType(Checkbox);
+      render(<CheckboxGroup options={options} />);
 
-      checkboxes.forEach((checkbox, index) => {
-        const option = options[index];
+      expect(renderMockCheckbox).toHaveBeenCalledTimes(options.length);
 
-        expect(checkbox.props.children).toBe(option.label);
-        expect(checkbox.props.disabled).toBe(option.disabled);
-        expect(checkbox.props.value).toBe(option.value);
+      options.forEach((opt) => {
+        expect(renderMockCheckbox).toHaveBeenCalledWith(
+          expect.objectContaining({
+            children: opt.label,
+            disabled: opt.disabled,
+            value: opt.value,
+          }),
+        );
       });
     });
 
@@ -116,79 +117,6 @@ describe('<CheckboxGroup />', () => {
       expect(firstElementChild!.getAttribute('data-test')).toBe('foo');
       expect(firstElementChild!.textContent).toBe('foo');
       expect(childElementCount).toBe(1);
-    });
-  });
-
-  describe('control', () => {
-    it('uncontrolled', () => {
-      const onChange = jest.fn();
-      const { getHostHTMLElement } = render(
-        <CheckboxGroup defaultValue={['bar']} onChange={onChange}>
-          <Checkbox value="foo">foo</Checkbox>
-          <Checkbox value="bar">bar</Checkbox>
-        </CheckboxGroup>,
-      );
-      const element = getHostHTMLElement();
-      const [fooCheckboxElement, barCheckboxElement] =
-        element.getElementsByTagName('input');
-
-      expect(fooCheckboxElement.checked).toBeFalsy();
-      expect(barCheckboxElement.checked).toBeTruthy();
-
-      fireEvent.click(barCheckboxElement);
-      expect(onChange).toBeCalledTimes(1);
-      expect(onChange.mock.calls[0][0]).toEqual([]);
-      expect(fooCheckboxElement.checked).toBeFalsy();
-      expect(barCheckboxElement.checked).toBeFalsy();
-
-      fireEvent.click(barCheckboxElement);
-      expect(onChange).toBeCalledTimes(2);
-      expect(onChange.mock.calls[1][0]).toEqual(['bar']);
-      expect(fooCheckboxElement.checked).toBeFalsy();
-      expect(barCheckboxElement.checked).toBeTruthy();
-
-      fireEvent.click(fooCheckboxElement);
-      expect(onChange).toBeCalledTimes(3);
-      expect(onChange.mock.calls[2][0]).toEqual(['bar', 'foo']);
-      expect(fooCheckboxElement.checked).toBeTruthy();
-      expect(barCheckboxElement.checked).toBeTruthy();
-    });
-
-    it('controlled', () => {
-      const TestingComponent = () => {
-        const [value, setValue] = useState(['bar']);
-
-        return (
-          <CheckboxGroup
-            defaultValue={['foo']}
-            onChange={setValue}
-            value={value}
-          >
-            <Checkbox value="foo">foo</Checkbox>
-            <Checkbox value="bar">bar</Checkbox>
-          </CheckboxGroup>
-        );
-      };
-
-      const { getHostHTMLElement } = render(<TestingComponent />);
-      const element = getHostHTMLElement();
-      const [fooCheckboxElement, barCheckboxElement] =
-        element.getElementsByTagName('input');
-
-      expect(fooCheckboxElement.checked).toBeFalsy();
-      expect(barCheckboxElement.checked).toBeTruthy();
-
-      fireEvent.click(barCheckboxElement);
-      expect(fooCheckboxElement.checked).toBeFalsy();
-      expect(barCheckboxElement.checked).toBeFalsy();
-
-      fireEvent.click(barCheckboxElement);
-      expect(fooCheckboxElement.checked).toBeFalsy();
-      expect(barCheckboxElement.checked).toBeTruthy();
-
-      fireEvent.click(fooCheckboxElement);
-      expect(fooCheckboxElement.checked).toBeTruthy();
-      expect(barCheckboxElement.checked).toBeTruthy();
     });
   });
 });
